pyproject.toml
[tool.poetry]
name = "megami2210"
version = "0.9.0"
description = "MEGUMI2210(TODO Application)"
authors = ["Teruaki.Sato"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
sqlalchemy = "^1.4.42"
psycopg2 = "^2.9.5"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.1"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"
pytest-cov = "^4.0.0"
pandas = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
